[ALPS04542773] DVFS: refine bit operation
correct data type for bit operations has been ported for
64 bit project inconsistency.

MTK-Commit-Id: 5c4eaabee1bb389affe018810a5e046349ad468b

CR-Id: ALPS04542773
Feature: DVFS (Dynamic Voltage and Frequency Scaling)

+++ drivers/misc/mediatek/base/power/cpufreq_v2/inc/mtk_cpufreq_internal.h
@@ -148,12 +148,12 @@ static const struct file_operations name ## _proc_fops = {		\
 /*
  * BIT Operation
  */
-#define _BIT_(_bit_)                    (unsigned long)(1 << (_bit_))
+#define _BIT_(_bit_)                    (unsigned int)(1 << (_bit_))
 #define _BITS_(_bits_, _val_) \
-	((((unsigned long) -1 >> (31 - ((1) ? _bits_))) & \
+	((((unsigned int) -1 >> (31 - ((1) ? _bits_))) & \
 	~((1U << ((0) ? _bits_)) - 1)) & ((_val_)<<((0) ? _bits_)))
 #define _BITMASK_(_bits_)               \
-	(((unsigned long) -1 >> (31 - ((1) ? _bits_))) & \
+	(((unsigned int) -1 >> (31 - ((1) ? _bits_))) & \
 	~((1U << ((0) ? _bits_)) - 1))
 #define _GET_BITS_VAL_(_bits_, _val_)   \
 	(((_val_) & (_BITMASK_(_bits_))) >> ((0) ? _bits_))
